Clean up FAST component usage to have a more reasonable structure
