+++ packages/site/src/fast/unwrapped/anchor.ts
@@ -1,0 +1,16 @@
+import { anchorTemplate, AnchorOptions } from '@microsoft/fast-foundation';
+import { Anchor, anchorStyles } from '@microsoft/fast-components';
+import { Constructable, css } from '@microsoft/fast-element';
+
+const fastAnchor = Anchor.compose<AnchorOptions, Constructable<Anchor>>({
+	baseName: 'anchor',
+	template: anchorTemplate,
+	styles: (ctx, def) => css`
+		${anchorStyles(ctx, def)}
+		.control {
+			align-items: center !important;
+		}
+	`,
+});
+
+export const anchor = fastAnchor();
